remove href
href tags were causing stuttering and jumping by forcing the page to reload. the consequence is loss of the navigation bar, but I don't think anyone was using that on PC. When it was working on mobile it was very buggy. An implementation where just some pages let you swipe back to drink list would be better